Calculator making
As part of the 100 Day Coding Challenge, I've developed a responsive calculator. This project not only offers a practical example of web development but also provides a valuable resource for users to practice and learn from.

Key Features:

Responsive Design: The calculator is designed to work seamlessly on various screen sizes, including mobile devices.

User-Friendly Interface: The calculator provides a simple and intuitive user interface, making it easy for users to perform mathematical calculations.

HTML, CSS, and JavaScript: I used a combination of HTML for structure, CSS for styling, and JavaScript for functionality. This stack is commonly used in web development and serves as an excellent learning opportunity for users